requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.18
djangorestframework==3.15.2
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
